refactor(amazonq): move src/inline -> src/app/inline
