+++ packages/amazonq/src/app/inline/activation.ts
