+++ packages/amazonq/src/app/inline/completion.ts
